+++ src/main/java/fll/web/api/AdvancingTeamsServlet.java
@@ -53,7 +53,8 @@ public class AdvancingTeamsServlet extends HttpServlet {
     final HttpSession session = request.getSession();
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.isHeadJudge()) {
+    if (!auth.isHeadJudge()
+        && !auth.isReportGenerator()) {
       response.sendError(HttpServletResponse.SC_FORBIDDEN);
       return;
     }
@@ -84,7 +85,8 @@ public class AdvancingTeamsServlet extends HttpServlet {
     final HttpSession session = request.getSession();
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.isHeadJudge()) {
+    if (!auth.isHeadJudge()
+        && !auth.isReportGenerator()) {
       response.sendError(HttpServletResponse.SC_FORBIDDEN);
       return;
     }