+++ src/main/java/fll/web/api/AwardGroupsServlet.java
@@ -43,7 +43,7 @@ public class AwardGroupsServlet extends HttpServlet {
     final HttpSession session = request.getSession();
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.isJudge()) {
+    if (!auth.isJudge() && !auth.isReportGenerator()) {
       response.sendError(HttpServletResponse.SC_FORBIDDEN);
       return;
     }