+++ src/main/java/fll/web/api/CategoryScheduleMappingServlet.java
@@ -45,7 +45,8 @@ public class CategoryScheduleMappingServlet extends HttpServlet {
     final HttpSession session = request.getSession();
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.isJudge()) {
+    if (!auth.isJudge()
+        && !auth.isReportGenerator()) {
       response.sendError(HttpServletResponse.SC_FORBIDDEN);
       return;
     }