+++ src/main/java/fll/web/report/PromptSummarizeScores.java
@@ -50,7 +50,7 @@ public class PromptSummarizeScores extends BaseFLLServlet {
       throws IOException, ServletException {
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.requireRoles(request, response, session, Set.of(UserRole.HEAD_JUDGE), false)) {
+    if (!auth.requireRoles(request, response, session, Set.of(UserRole.HEAD_JUDGE, UserRole.REPORT_GENERATOR), false)) {
       return;
     }
 
@@ -91,7 +91,8 @@ public class PromptSummarizeScores extends BaseFLLServlet {
     }
 
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
-    if (!auth.isHeadJudge()) {
+    if (!auth.isHeadJudge()
+        && !auth.isReportGenerator()) {
       // only the head judge can summarize scores, so don't prompt others to summarize
       // the scores
       return false;