+++ src/main/java/fll/web/report/SubjectiveByJudge.java
@@ -70,7 +70,7 @@ public class SubjectiveByJudge extends BaseFLLServlet {
 
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.requireRoles(request, response, session, Set.of(UserRole.JUDGE), false)) {
+    if (!auth.requireRoles(request, response, session, Set.of(UserRole.JUDGE, UserRole.REPORT_GENERATOR), false)) {
       return;
     }